refactor: Move selection types to `App.Base`
This will be necessary in order to make use of them in actions-related modules.
